Traps!
Trap script, select the interval when it will set in the inspector.
Select the damage it does. (default 33)
White when inactive
Red when active.
It shoots a raycast from above the trap and goes down towards where the player should be. If it detects the player it lowers the soup value.